+++ .travis.yml
@@ -2,4 +2,4 @@ language: python
 python:
  - "3.7-dev"
 script:
-  - python3 -m unittest discover cli/tests/
+ - pytest
