+++ cli/tests/example.txt
