+++ cli/tests/test_command_cat.py
@@ -1,14 +1,14 @@
 from unittest import TestCase
 
-from commands import CommandCat
-from storage import Storage
+from src.commands import CommandCat
+from src.storage import Storage
 
 
 class TestCommandCat(TestCase):
     def test_execute(self):
         storage = Storage(r'\$[^ \'\"$]+')
 
-        command = CommandCat(['../example.txt'])
+        command = CommandCat(['example.txt'])
         self.assertEqual(command.execute("", storage), "Some example text")
 
         command = CommandCat(['dsakfjhakdsljf'])
