+++ cli/tests/test_command_wc.py
@@ -1,14 +1,14 @@
 from unittest import TestCase
 
-from commands import CommandWC
-from storage import Storage
+from src.commands import CommandWC
+from src.storage import Storage
 
 
 class TestCommandWC(TestCase):
     def test_execute(self):
         storage = Storage(r'\$[^ \'\"$]+')
 
-        command = CommandWC(['../example.txt'])
+        command = CommandWC(['example.txt'])
         self.assertEqual(command.execute("", storage), "1 3 17")
 
         command = CommandWC([])
